pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.7"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
gunicorn = "^21.2.0"
django-widget-tweaks = "^1.4.12"
pillow = "^10.0.0"
numpy = "^1.25.2"
django-admin = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
